Fix windows taskbar notifications counter with multi-windows.
