+++ Telegram/SourceFiles/platform/win/integration_win.h
@@ -7,9 +7,41 @@ https://github.com/telegramdesktop/tdesktop/blob/master/LEGAL
 */
 #pragma once
 
+#include "platform/platform_integration.h"
+
+#include <QAbstractNativeEventFilter>
+
+#include <winrt/base.h>
+#include <ShlObj.h>
+
 namespace Platform {
 
-class Integration;
+class WindowsIntegration final
+	: public Integration
+	, public QAbstractNativeEventFilter {
+public:
+	void init() override;
+
+	[[nodiscard]] ITaskbarList3 *taskbarList() const;
+
+	[[nodiscard]] static WindowsIntegration &Instance();
+
+private:
+	bool nativeEventFilter(
+		const QByteArray &eventType,
+		void *message,
+		long *result) override;
+	bool processEvent(
+		HWND hWnd,
+		UINT msg,
+		WPARAM wParam,
+		LPARAM lParam,
+		LRESULT *result);
+
+	uint32 _taskbarCreatedMsgId = 0;
+	winrt::com_ptr<ITaskbarList3> _taskbarList;
+
+};
 
 [[nodiscard]] std::unique_ptr<Integration> CreateIntegration();
